+++ src/FormLabel/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as FormLabel } from './FormLabel';
