+++ src/TextInput/TextInput.js
@@ -11,7 +11,7 @@ import {
 import { defaultTokens } from '@kiwicom/orbit-design-tokens';
 
 import { Text } from '../Text';
-import FormLabel from './FormLabel';
+import { FormLabel } from '../FormLabel';
 import { Icon } from '../Icon';
 import { FormFeedback } from '../FormFeedback';
 import { StyleSheet } from '../PlatformStyleSheet';
